+++ camouchat/WhatsApp/core/login.py
@@ -15,7 +15,7 @@ from playwright.async_api import (
 )
 from typing import Optional, Union
 from camouchat.Exceptions.whatsapp import LoginError
-from camouchat.Interfaces.login_interface import LoginInterface
+from camouchat.contracts.login_interface import LoginInterface
 from camouchat.WhatsApp.core.web_ui_config import WebSelectorConfig
 from camouchat.camouchat_logger import camouchatLogger